chore: cache: move `baseClient` functionality to Memcached client
Remove the `baseClient` that existed to share functionality between
Redis and Memcached clients now that the Redis client has been removed.
